requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cassandra-driver==3.28.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cssselect==1.2.0
exceptiongroup==1.1.2
fake-useragent==1.1.3
fastapi==0.100.0
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
outcome==1.2.0
parse==1.19.1
pydantic==2.0.3
pydantic_core==2.3.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-html==0.10.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.1
w3lib==2.1.1
websockets==10.4
wsproto==1.2.0
zipp==3.16.2
